Fixed chord normalisation code, should work now

+++ app/src/main/java/com/example/piano/KeyEvent.java
@@ -3,10 +3,10 @@ package com.example.piano;
 public class KeyEvent{
 
     public int note;
-    public float time;
+    public long time;
     public boolean press;
 
-    public KeyEvent(int note, float time, boolean press) {
+    public KeyEvent(int note, long time, boolean press) {
         this.note = note;
         this.time = time;
         this.press = press;
